requirements.txt
dotenv==0.9.9
pillow==11.2.1
pyodbc==5.2.0
python-dotenv==1.1.0
